// pong/src/utils/Types.ts
export interface DepthLineBehavior {
    movement?: 'vertical' | 'horizontal' | string;
    direction?: 'upwards' | 'downwards' | 'left' | 'right' | string;
    fade?: 'in' | 'out' | 'none' | string;
    
	// Pyramid specific properties
    pyramidBaseHeight?: number;
    pyramidBaseWidth?: number;
    pyramidPeakHeight?: number;
    pyramidPeakOffset?: number;

	// Ruin specicif properties
	ruinHSegments?: number;
	ruinTSegments?: number;
	maxHeight?: number;
	maxWidth?: number;
	segmentWidths?: number[];
    segmentHeights?: number[];
	hOffset?: number;

	// Wave specific properties
    waveAmplitude?: number;
    waveFrequency?: number;
    waveOffset?: number;
    phaseShift?: number;
    waveType?: 'sine' | 'square' | 'triangle' | 'sawtooth';
    amplitudeVariation?: number;   // Add this property
    minBottomHeight?: number | null;  // Add this property
}